+++ retz-server/src/test/java/io/github/retz/scheduler/RetzSchedulerTest.java
@@ -95,7 +95,7 @@ public class RetzSchedulerTest {
     //TODO: @Test
     public void launch() throws InterruptedException, JsonProcessingException, IOException {
         String files[] = {"http://foobar.boom.co.jp/foo.tar.gz"};
-        Applications.load(new Application("fooapp", new LinkedList<String>(), Arrays.asList(files), Optional.empty()));
+        Applications.load(new Application("fooapp", new LinkedList<String>(), new LinkedList<String>(), Arrays.asList(files), Optional.empty()));
         Job job = new Job("fooapp", "foocmd", null, new Range(2, 0), new Range(256, 0));
         JobQueue.push(job);
 
@@ -137,7 +137,7 @@ public class RetzSchedulerTest {
     @Test
     public void notEnough() throws InterruptedException {
         String files[] = {"http://foobar.boom.co.jp/foo.tar.gz"};
-        Applications.load(new Application("fooapp", new LinkedList<String>(), Arrays.asList(files), Optional.empty()));
+        Applications.load(new Application("fooapp", new LinkedList<String>(), new LinkedList<String>(), Arrays.asList(files), Optional.empty()));
         Job job = new Job("fooapp", "foocmd", null, new Range(2, 0), new Range(256, 0));
         JobQueue.push(job);
 
@@ -160,7 +160,7 @@ public class RetzSchedulerTest {
     // TODO: @Test
     public void slaveFail() throws InterruptedException, JsonProcessingException, IOException {
         String files[] = {"http://foobar.boom.co.jp/foo.tar.gz"};
-        Applications.load(new Application("fooapp", new LinkedList<String>(), Arrays.asList(files), Optional.empty()));
+        Applications.load(new Application("fooapp", new LinkedList<String>(), new LinkedList<String>(), Arrays.asList(files), Optional.empty()));
         Job job = new Job("fooapp", "foocmd", null, new Range(2, 0), new Range(256, 0));
         JobQueue.push(job);